Separate the Configuration API into submodules
